Update docker-hotreload.sh
Add optional -s argument flag for stop_cmd to quite out the process started by the build_cmd. It also ensures if the build_cmd is provided and is not a docker or docker-compose command, and the stop_cmd was not provided that we pass an error to the user and exit.